Fix minor compiler warnings

+++ code/game/q_shared.h
@@ -486,7 +486,7 @@ typedef enum
 
 q_static_assert(NUM_FORCE_POWERS < 31);
 
-#define FP_Selectable(x) ((x) >= FP_FIRST && (x) < FP_SABERATTACK && (x) != FP_LEVITATION)
+#define FP_Selectable(x) ((int)(x) >= FP_FIRST && (x) < FP_SABERATTACK && (x) != FP_LEVITATION)
 
 typedef enum
 {
